requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
DBUtils==1.3
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
typing==3.7.4.1
urllib3==1.25.8
Werkzeug==1.0.0
xmltodict==0.12.0
